% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{CDFApproximation}
\alias{CDFApproximation}
\title{CDF approximation.}
\usage{
CDFApproximation(fractions, quantiles)
}
\arguments{
\item{fractions}{Quantile fractions.}

\item{quantiles}{Quantiles.}
}
\description{
Computes an approximated PDF function
using vectors of quantile fractions and quantiles.
}
\seealso{
Other Distribution functions: \code{\link{ExpectedValueApproximation}},
  \code{\link{PDFApproximation}},
  \code{\link{QRMonConditionalCDF}}
}
\concept{Distribution functions}
