% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{CDFRandomPoints}
\alias{CDFRandomPoints}
\title{Random points generation for a CDF.}
\usage{
CDFRandomPoints(cdf, n = 1, randomFunction = runif)
}
\arguments{
\item{cdf}{A data frame with numeric columns "Probability" and "Quantile".}

\item{n}{Number of points.}

\item{randomFunction}{A function to generate a random point between the quantiles.
(It rarely makes sense to be different than \code{runif}.)}
}
\value{
A numeric vector
}
\description{
Generates a vector of random points using a CDF data frame.
}
