% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{ExpectedValueApproximation}
\alias{ExpectedValueApproximation}
\title{Expected value approximation.}
\usage{
ExpectedValueApproximation(fractions, quantiles)
}
\arguments{
\item{fractions}{Quantile fractions.}

\item{quantiles}{Quantile values.}
}
\description{
Computes an approximated expected value
using vectors of quantiles and quantile values.
}
\seealso{
Other Distribution functions: \code{\link{CDFApproximation}},
  \code{\link{PDFApproximation}},
  \code{\link{QRMonConditionalCDF}}
}
\concept{Distribution functions}
