% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{PDFApproximation}
\alias{PDFApproximation}
\title{PDF approximation.}
\usage{
PDFApproximation(probabilities, quantiles)
}
\arguments{
\item{probabilities}{Quantile probabilities.}

\item{quantiles}{Quantile values.}
}
\description{
Computes an approximated PDF function
using vectors of quantiles and quantile values.
}
\seealso{
Other Distribution functions: 
\code{\link{CDFApproximation}()},
\code{\link{ExpectedValueApproximation}()},
\code{\link{QRMonConditionalCDFPlot}()},
\code{\link{QRMonConditionalCDF}()}
}
\concept{Distribution functions}
