% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonConditionalCDF}
\alias{QRMonConditionalCDF}
\title{Conditional CDF computation.}
\usage{
QRMonConditionalCDF(qrObj, timePoints)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{timePoints}{Time points to compute CDF's upon.}
}
\value{
A QRMon object.
}
\description{
Computes conditional CDF's for given time points.
}
\details{
The computations result is assigned to \code{qrObj$Value}.
}
