% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonConditionalCDFPlot}
\alias{QRMonConditionalCDFPlot}
\title{Conditional CDF plot.}
\usage{
QRMonConditionalCDFPlot(qrObj, regressorValue, valueGridPoints = NULL,
  echoQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{regressorValue}{A regressor point to compute CDF's upon.}

\item{valueGridPoints}{Grid points to use for the value(response) variable.
If NULL the grid points are derived from response variable's range in the data.}

\item{echoQ}{To echo the plot the or not?}
}
\value{
A QRMon object.
}
\description{
Computes conditional CDF's for given regressor points and makes corresponding plots.
}
\details{
This function uses \link{\code{QRMonConditionalCDF}}.
}
\seealso{
Other Distribution functions: \code{\link{CDFApproximation}},
  \code{\link{ExpectedValueApproximation}},
  \code{\link{PDFApproximation}},
  \code{\link{QRMonConditionalCDF}}
}
\concept{Distribution functions}
