% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonDataCheck}
\alias{QRMonDataCheck}
\title{Check presence of required data.}
\usage{
QRMonDataCheck(qrObj, functionName = NULL, logicalResult = FALSE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{functionName}{A name of the delegating function (if any).}

\item{logicalResult}{Should the result be logical value?}
}
\value{
If \code{logicalValue} is FALSE the result is QRMon object; if TRUE the result is logical value.
}
\description{
Checks does an QRMon object have event records, entity attributes, and computation specification.
}
