% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonDeleteMissing}
\alias{QRMonDeleteMissing}
\title{Remove missing data.}
\usage{
QRMonDeleteMissing(qrObj)
}
\arguments{
\item{qrObj}{An QRMon object.}
}
\value{
A QRMon object.
}
\description{
Removes data rows with missing data fields (NA's).
}
\details{
This function is most likely reduntant.
The function \code{\link{QRMonSetData}} does removal of missing data.
}
