% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonEcho}
\alias{QRMonEcho}
\title{Echo.}
\usage{
QRMonEcho(qrObj, x)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{x}{An object to be echoed.}
}
\value{
A QRMon object.
}
\description{
Echoes given object.
}
\seealso{
Other Echo functions: 
\code{\link{QRMonEchoFunctionValue}()},
\code{\link{QRMonEchoValue}()}
}
\concept{Echo functions}
