% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonEchoDataSummary}
\alias{QRMonEchoDataSummary}
\title{Data summary.}
\usage{
QRMonEchoDataSummary(qrObj)
}
\arguments{
\item{qrObj}{An QRMon object.}
}
\value{
A QRMon object.
}
\description{
Summarize data.
using a B-spline basis.
}
\details{
Prints data dimensions and summary.
}
