% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonEchoFunctionValue}
\alias{QRMonEchoFunctionValue}
\title{Function application to monad's value.}
\usage{
QRMonEchoFunctionValue(qrObj, f)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{f}{A function to be applied to \code{qrObj$Value}.}
}
\value{
A QRMon object.
}
\description{
Apply a function to the "Value" element/member of the monad object.
}
\details{
Prints \code{f(qrObj$Value)}.
}
