% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonEchoValue}
\alias{QRMonEchoValue}
\title{Echo monad's value.}
\usage{
QRMonEchoValue(qrObj)
}
\arguments{
\item{qrObj}{An QRMon object.}
}
\value{
A QRMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{qrObj$Value}.
}
