% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonErrors}
\alias{QRMonErrors}
\title{Fit errors.}
\usage{
QRMonErrors(qrObj, relativeErrorsQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{relativeErrorsQ}{Should relative errors be computed?}
}
\value{
A QRMon object.
}
\description{
Find the fit errors of the monad object data and found regression objects.
}
\details{
The errors are assigned to \code{qrObj$Value}.
}
\seealso{
Other Errors: \code{\link{QRMonErrorsPlot}}
}
\concept{Errors}
