% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonErrorsPlot}
\alias{QRMonErrorsPlot}
\title{Fit errors plot.}
\usage{
QRMonErrorsPlot(
  qrObj,
  relativeErrorsQ = TRUE,
  datePlotQ = FALSE,
  dateOrigin = "1970-01-01",
  echoQ = TRUE
)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{relativeErrorsQ}{Should relative errors be computed?}

\item{datePlotQ}{Should the regressor axis have dates scale?}

\item{dateOrigin}{Same as the argument \code{origin} of \code{as.POSIXct}.}

\item{echoQ}{To echo the plot the or not?}
}
\value{
A QRMon object.
}
\description{
Find and plot the fit errors of the monad object data and found regression objects.
}
\details{
The errors plot are assigned to \code{qrObj$Value}.
}
\seealso{
Other Errors: 
\code{\link{QRMonErrors}()}
}
\concept{Errors}
