% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonFailure}
\alias{QRMonFailure}
\title{Failure producing function}
\usage{
QRMonFailure(msObj, message = NULL)
}
\arguments{
\item{msObj}{An QRMon object.}

\item{message}{A message to echoed. If NULL no message is echoed.}
}
\value{
A QRMon object.
}
\description{
Puts the monad object into a monadic failure.
}
