% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\docType{data}
\name{QRMonFailureSymbol}
\alias{QRMonFailureSymbol}
\title{Failure symbol for QRMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
QRMonFailureSymbol
}
\description{
Failure symbol for the monad QRMon.
}
\keyword{datasets}
