% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonFindAnomaliesByResiduals}
\alias{QRMonFindAnomaliesByResiduals}
\title{Find anomalies by residuals.}
\usage{
QRMonFindAnomaliesByResiduals(
  qrObj,
  threshold = NULL,
  outlierIdentifier = NULL,
  relativeErrorsQ = FALSE
)
}
\arguments{
\item{qrObj}{A QRMon object.}

\item{threshold}{The threshold to be used to identify anomalies.
If NULL the outlier identifier is used.}

\item{outlierIdentifier}{Outlier identifier of list of numbers
to be used to identify the anomalies.
If NULL an internal implementation of the quartile outlier
identifier is used.}

\item{relativeErrorsQ}{Should relative errors be used or not?}
}
\value{
A QRMon object.
}
\description{
Finds the anomalies of the data by using a residuals to
the fitted regression quantile and threshold or an outlier identifier.
}
\details{
The residuals outliers are picked with the formulas:
\code{c(xL,x0,xU) <- quantile( abs(residuals), c(1/4,1/2,3/4) )},
\code{abs(residuals) > x0+(xU-xL)}.
(Errors and residuals are used as synonyms here.)
}
