% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonFindVarianceAnomalies}
\alias{QRMonFindVarianceAnomalies}
\title{Find variance anomalies.}
\usage{
QRMonFindVarianceAnomalies(qrObj, outlierIdentifier = NULL, positionsQ = FALSE)
}
\arguments{
\item{qrObj}{A QRMon object.}

\item{outlierIdentifier}{Outlier identifier of list of numbers
to be used to identify the anomalies.
If NULL an internal implementation of the quartile outlier
identifier is used.}

\item{positionsQ}{Should position be returned or points?}
}
\value{
A QRMon object.
}
\description{
Finds the variance anomalies of the data by using
the differences of the fitted regression quantiles corresponding to
the lowest and highest probabilities and
an outlier identifier applied to those differences.
}
