% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonMemberPresenceCheck}
\alias{QRMonMemberPresenceCheck}
\title{General member presence check.}
\usage{
QRMonMemberPresenceCheck(qrObj, memberName,
  memberPrettyName = memberName, functionName = "",
  logicalResult = FALSE, warningQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{memberName}{The name of the member to be checked.}

\item{memberPrettyName}{A pretty member name (for messages).}

\item{functionName}{The name of the delegating function.}

\item{logicalResult}{Should the result be logical value?}

\item{warningQ}{Should a warning be issued if the specified member is not found?}
}
\value{
A logical value or an QRMon object.
}
\description{
A general function for checking the presence of a data member in an QRMon object.
}
