% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonOutliers}
\alias{QRMonOutliers}
\title{Outliers finding.}
\usage{
QRMonOutliers(qrObj)
}
\arguments{
\item{qrObj}{An QRMon object.}
}
\value{
A QRMon object.
}
\description{
Find the monad object data outliers using already found regression objects.
}
\details{
The outliers are assigned to \code{qrObj$Value} and \code{qrObj$Outliers}.
}
\seealso{
Other Outlier functions: \code{\link{QRMonOutliersPlot}}
}
\concept{Outlier functions}
