% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonOutliersPlot}
\alias{QRMonOutliersPlot}
\title{Outliers plot.}
\usage{
QRMonOutliersPlot(qrObj, plotRegressionCurvesQ = TRUE, echoQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{plotRegressionCurvesQ}{Should the regression curves be plotted or not?}

\item{echoQ}{To echo the plot the or not?}
}
\value{
A QRMon object.
}
\description{
Plot the monad object data and found outliers.
}
\details{
The outliers are assigned to \code{qrObj$Value} and \code{qrObj$Outliers}.
}
\seealso{
Other Outliers: \code{\link{QRMonOutliers}}
}
\concept{Outliers}
