% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonPickPathPoints}
\alias{QRMonPickPathPoints}
\title{Pick path points.}
\usage{
QRMonPickPathPoints(qrObj, threshold, pickAboveThresholdQ = FALSE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{threshold}{The pick threshold.}

\item{pickAboveThresholdQ}{Should points be picked below or above the threshold?}
}
\value{
A QRMon object.
}
\description{
Pick points close to the regression functions using a specified threshold.
}
\details{
The list of data frames is assigned to \code{qrObj$Value}.
}
\seealso{
Other Regression functions: \code{\link{QRMonPredict}},
  \code{\link{QRMonQuantileRegression}}
}
\concept{Regression functions}
