% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonPlot}
\alias{QRMonPlot}
\title{Plot with data and regression curves.}
\usage{
QRMonPlot(qrObj, echoQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{echoQ}{To echo the plot the or not.}
}
\value{
A QRMon object.
}
\description{
Plot the monad object data and regression functions (if any.)
}
\details{
The plot is made with \link{ggplot2}.
The plot is assigned to \code{qrObj$Value}.
}
\concept{Plots}
