% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonQuantileRegression}
\alias{QRMonQuantileRegression}
\title{Quantile regression B-splines basis fit.}
\usage{
QRMonQuantileRegression(qrObj, fractions = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{fractions}{A numeric vector with quantile fractions.}

\item{...}{parameters for \code{\link{splines::bs}}.}
}
\value{
A QRMon object.
}
\description{
Finds the quantile regression objects for the specified
quantile fractions using a B-spline basis.
}
\details{
The obtained regression objects are assigned/appended to the
\code{qrObj$RegressionObjects}.
For more computational details see \code{\link{quantreg::rq}}.
This function can be seen as a splines-basis shortcut of
\code{\link{QRMonQuantileRegressionFit}}.
}
\seealso{
Other Regression functions: \code{\link{QRMonPickPathPoints}},
  \code{\link{QRMonPredict}},
  \code{\link{QRMonQuantileRegressionFit}},
  \code{\link{QRMonSeparateToFractions}},
  \code{\link{QRMonSeparate}}
}
\concept{Regression functions}
