% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSeparate}
\alias{QRMonSeparate}
\title{Separate data points.}
\usage{
QRMonSeparate(qrObj, data = NULL, cumulativeQ = TRUE,
  fractionsQ = FALSE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{data}{A data frame of points to be separated.}

\item{cumulativeQ}{Should for each regression quantile find the points below it?}

\item{fractionsQ}{Should fractions instead of points be returned?}
}
\value{
A QRMon object.
}
\description{
Separates the argument by the regression functions in the context.
If no argument is given the data in the monad object is separated.
}
\details{
The result of the separation is a list of data frames assigned to \code{qrObj$Value}.
Each data frame of that list corresponds to the found regression quantiles.
If the data frame argument \code{data} has columns "Regressor" and "Value" those columns are used;
otherwise the first and second columns are treated as "Regressor" and "Value" respectively.
}
\seealso{
Other Regression functions: \code{\link{QRMonPickPathPoints}},
  \code{\link{QRMonPredict}},
  \code{\link{QRMonQuantileRegressionFit}},
  \code{\link{QRMonQuantileRegression}},
  \code{\link{QRMonSeparateToFractions}}
}
\concept{Regression functions}
