% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSeparateToFractions}
\alias{QRMonSeparateToFractions}
\title{Fractions of separated data points.}
\usage{
QRMonSeparateToFractions(qrObj, data = NULL, cumulativeQ = TRUE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{data}{A data frame of points to be separated.}

\item{cumulativeQ}{Should for each regression quantile find the points below it?}

\item{fractionsQ}{Should fractions instead of points be returned?}
}
\value{
A QRMon object.
}
\description{
Separates the argument by the regression functions in the context
and compute the corresponding fractions.
If no argument is given the data in the monad object is separated.
}
\details{
The result is assigned to \code{qrObj$Value}.
For more details see \code{\link{QRMonSeparate}}.
(This function is just a shortcut to that one.)
}
\seealso{
Other Regression functions: 
\code{\link{QRMonPickPathPoints}()},
\code{\link{QRMonPredict}()},
\code{\link{QRMonQuantileRegressionFit}()},
\code{\link{QRMonQuantileRegression}()},
\code{\link{QRMonSeparate}()}
}
\concept{Regression functions}
