% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSetData}
\alias{QRMonSetData}
\title{Set event records.}
\usage{
QRMonSetData(qrObj, data)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{data}{A data frame with event records.}
}
\value{
An QRMon object.
}
\description{
Sets an event records data frame into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{QRMonSetRegressionObjects}()},
\code{\link{QRMonSetValue}()},
\code{\link{QRMonTakeData}()},
\code{\link{QRMonTakeOutliers}()},
\code{\link{QRMonTakeRegressionObjects}()},
\code{\link{QRMonTakeValue}()}
}
\concept{Set/Take functions}
