% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSetRegressionObjects}
\alias{QRMonSetRegressionObjects}
\title{Set regression functions.}
\usage{
QRMonSetRegressionObjects(qrObj, regressionObjects)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{regressionObjects}{A list of regression objects.}
}
\value{
An QRMon object.
}
\description{
Sets a list of regression functions into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{QRMonSetData}},
  \code{\link{QRMonTakeData}},
  \code{\link{QRMonTakeOutliers}},
  \code{\link{QRMonTakeRegressionObjects}},
  \code{\link{QRMonTakeValue}}
}
\concept{Set/Take functions}
