% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSetValue}
\alias{QRMonSetValue}
\title{Set the value in a QRMon object.}
\usage{
QRMonSetValue(qrObj, value)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{value}{The new value.}
}
\value{
A QRMon object.
}
\description{
Sets the value in a QRMon monad object.
}
\details{
Assigns \code{value} to \code{qrObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{QRMonSetData}()},
\code{\link{QRMonSetRegressionObjects}()},
\code{\link{QRMonTakeData}()},
\code{\link{QRMonTakeOutliers}()},
\code{\link{QRMonTakeRegressionObjects}()},
\code{\link{QRMonTakeValue}()}
}
\concept{Set/Take functions}
