% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSimulate}
\alias{QRMonSimulate}
\title{Simulate data points.}
\usage{
QRMonSimulate(qrObj, n = 100)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{n}{Number of simulated points.}
}
\value{
A QRMon object.
}
\description{
Simulate data points based on regression quantiles.
}
\details{
The data frame with the simulated points are assigned to \code{qrObj$Value}.
}
