% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonTakeData}
\alias{QRMonTakeData}
\title{Take event records.}
\usage{
QRMonTakeData(qrObj, functionName = "QMonTakeData")
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A data frame or \code{QRMonFailureSymbol}.
}
\description{
Takes the event records data frame from the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{QRMonSetData}},
  \code{\link{QRMonSetRegressionObjects}},
  \code{\link{QRMonTakeRegressionObjects}},
  \code{\link{QRMonTakeValue}}
}
\concept{Set/Take functions}
