% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonTakeOutliers}
\alias{QRMonTakeOutliers}
\title{Take outliers.}
\usage{
QRMonTakeOutliers(qrObj, functionName = "QRMonTakeOutliers")
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{QRMonFailureSymbol}.
}
\description{
Takes the outliers from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{QRMonSetData}()},
\code{\link{QRMonSetRegressionObjects}()},
\code{\link{QRMonSetValue}()},
\code{\link{QRMonTakeData}()},
\code{\link{QRMonTakeRegressionObjects}()},
\code{\link{QRMonTakeValue}()}
}
\concept{Set/Take functions}
