% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonTakeRegressionObjects}
\alias{QRMonTakeRegressionObjects}
\title{Take regression functions.}
\usage{
QRMonTakeRegressionObjects(qrObj,
  functionName = "QRMonTakeRegressionObjects")
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{QRMonFailureSymbol}.
}
\description{
Takes the regression functions from the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{QRMonSetData}},
  \code{\link{QRMonSetRegressionObjects}},
  \code{\link{QRMonTakeData}}, \code{\link{QRMonTakeValue}}
}
\concept{Set/Take functions}
