% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonTakeValue}
\alias{QRMonTakeValue}
\title{Take the value in a QRMon object.}
\usage{
QRMonTakeValue(qrObj)
}
\arguments{
\item{qrObj}{An QRMon object.}
}
\value{
Just \code{qrObj$Value}.
}
\description{
Takes the value from QRMon monad object.
}
\seealso{
Other Set/Take functions: \code{\link{QRMonSetData}},
  \code{\link{QRMonSetRegressionObjects}},
  \code{\link{QRMonTakeData}},
  \code{\link{QRMonTakeOutliers}},
  \code{\link{QRMonTakeRegressionObjects}}
}
\concept{Set/Take functions}
