% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonUnit}
\alias{QRMonUnit}
\title{Make a QRMon Unit}
\usage{
QRMonUnit(data = NULL)
}
\arguments{
\item{data}{A vector or a two-column matrix or data frame.}
}
\value{
An S3 class "QRMon". In other words, a list with the attribute "class" set to "QRMon".
}
\description{
Creates a monad object.
}
