% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{CDFApproximation}
\alias{CDFApproximation}
\title{CDF approximation.}
\usage{
CDFApproximation(qs, qvals)
}
\arguments{
\item{qs}{Quantiles.}

\item{qvals}{Quantile values.}
}
\description{
Computes an approximated PDF function
using vectors of quantiles and quantile values.
}
\seealso{
Other Distribution functions: \code{\link{ExpectedValueApproximation}},
  \code{\link{PDFApproximation}},
  \code{\link{QRMonConditionalCDF}}
}
\concept{Distribution functions}
