% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{CDFApproximation}
\alias{CDFApproximation}
\title{CDF approximation.}
\usage{
CDFApproximation(probabilities, quantiles)
}
\arguments{
\item{probabilities}{Quantile probabilities.}

\item{quantiles}{Quantiles.}
}
\description{
Computes an approximated PDF function
using vectors of quantile probabilities and quantiles.
}
\seealso{
Other Distribution functions: 
\code{\link{ExpectedValueApproximation}()},
\code{\link{PDFApproximation}()},
\code{\link{QRMonConditionalCDFPlot}()},
\code{\link{QRMonConditionalCDF}()}
}
\concept{Distribution functions}
