% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonConditionalCDFPlot}
\alias{QRMonConditionalCDFPlot}
\title{Conditional CDF plot.}
\usage{
QRMonConditionalCDFPlot(qrObj, regressorValues, valueGridPoints = NULL,
  dateOrigin = NULL, quantileGridLinesQ = TRUE, echoQ = TRUE, ...)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{regressorValues}{Regressor points to compute CDF's upon.}

\item{valueGridPoints}{Grid points to use for the value(response) variable.
If NULL the grid points are derived from response variable's range in the data.}

\item{dateOrigin}{Date origin if the regressor conversiont to time-date.
If NULL no conversion is done.
Same as the argument \code{origin} of \code{as.POSIXct}.}

\item{quantileGridLinesQ}{Should the quantiles be indicated with vertical grid lines or not?}

\item{echoQ}{To echo the plot the or not?}

\item{...}{Arguments for \link{\code{ggplot2::facet_wrap}}.
(In order to plot additional vertical lines.)}
}
\value{
A QRMon object.
}
\description{
Computes conditional CDF's for given regressor points and makes corresponding plots.
}
\details{
This function uses \link{\code{QRMonConditionalCDF}}.
}
\seealso{
Other Distribution functions: \code{\link{CDFApproximation}},
  \code{\link{ExpectedValueApproximation}},
  \code{\link{PDFApproximation}},
  \code{\link{QRMonConditionalCDF}}
}
\concept{Distribution functions}
