% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonOutliersPlot}
\alias{QRMonOutliersPlot}
\title{Outliers plot.}
\usage{
QRMonOutliersPlot(
  qrObj,
  plotRegressionCurvesQ = TRUE,
  dataLineColor = NULL,
  dataPointsColor = "gray60",
  datePlotQ = FALSE,
  dateOrigin = "1970-01-01",
  echoQ = TRUE
)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{plotRegressionCurvesQ}{Should the regression curves be plotted or not?}

\item{dataLineColor}{The color of the line connecting the data points.
If NULL the data line is not plotted.}

\item{dataPointsColor}{The color of the data points.
If NULL the data points are not plotted.}

\item{datePlotQ}{Should the regressor axis have dates scale?}

\item{dateOrigin}{Same as the argument \code{origin} of \code{as.POSIXct}.}

\item{echoQ}{To echo the plot the or not?}
}
\value{
A QRMon object.
}
\description{
Plot the monad object data and found outliers.
}
\details{
The outliers are assigned to \code{qrObj$Value} and \code{qrObj$Outliers}.
}
\seealso{
Other Outlier functions: 
\code{\link{QRMonOutliers}()}
}
\concept{Outlier functions}
