% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonQuantileRegression}
\alias{QRMonQuantileRegression}
\title{Quantile regression B-splines basis fit.}
\usage{
QRMonQuantileRegression(qrObj, df, knots = NULL, degree = 3,
  probabilities = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{df}{Degrees of freedom;
same as \code{df} of \code{\link{splines::bs}}.}

\item{knots}{The internal breakpoints that define the spline;
same as \code{knots} of \code{\link{splines::bs}}.}

\item{degree}{Degree of the piecewise polynomial;
same as \code{degree} of \code{\link{splines::bs}}.}

\item{probabilities}{A numeric vector with quantile probabilities.}

\item{...}{Additional arguments for \code{\link{splines::bs}}.}
}
\value{
A QRMon object.
}
\description{
Finds the quantile regression objects for the specified
quantile probabilities using a B-spline basis.
}
\details{
The obtained regression objects are assigned/appended to the
\code{qrObj$RegressionObjects}.
For more computational details see \code{\link{quantreg::rq}}.
This function can be seen as a splines-basis shortcut of
\code{\link{QRMonQuantileRegressionFit}}.
}
\seealso{
Other Regression functions: \code{\link{QRMonPickPathPoints}},
  \code{\link{QRMonPredict}},
  \code{\link{QRMonQuantileRegressionFit}},
  \code{\link{QRMonSeparateToFractions}},
  \code{\link{QRMonSeparate}}
}
\concept{Regression functions}
