% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonQuantileRegression}
\alias{QRMonQuantileRegression}
\title{Quantile regression fit≈.}
\usage{
QRMonQuantileRegression(qrObj, quantiles = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{quantiles}{A numeric vector with quantiles.}

\item{...}{parameters for \code{\link{quantreg::rq}}.}
}
\value{
A QRMon object.
}
\description{
Finds the quantile regression objects for the specified quantiles
using a B-spline basis.
}
\details{
The obtained regression objects are assigned/appended to the
\code{qrObj$RegressionObjects}.
}
\seealso{
Other Regression functions: \code{\link{QRMonPickPathPoints}},
  \code{\link{QRMonPredict}}
}
\concept{Regression functions}
