% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonQuantileRegressionFit}
\alias{QRMonQuantileRegressionFit}
\title{Quantile regression function basis fit.}
\usage{
QRMonQuantileRegressionFit(qrObj, functionBasis, quantiles = c(0.25, 0.5,
  0.75))
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{functionBasis}{A list of basis functions or
a basis matrix computed over the data abscissas.}

\item{quantiles}{A numeric vector with quantiles.}
}
\value{
A QRMon object.
}
\description{
Finds the quantile regression objects for the specified quantiles
using a specified function basis.
}
\details{
When the argument \code{functionBasis} is a (basis) matrix it is expected
to be evaluated over the data abscissas (\code{qrObj$Data$Time}).
The obtained regression objects are assigned/appended to the
\code{qrObj$RegressionObjects}.
For more computational details see \code{\link{quantreg::rq}}.
}
\seealso{
Other Regression functions: \code{\link{QRMonPickPathPoints}},
  \code{\link{QRMonPredict}},
  \code{\link{QRMonQuantileRegression}},
  \code{\link{QRMonSeparateToFractions}},
  \code{\link{QRMonSeparate}}
}
\concept{Regression functions}
