% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonRescale}
\alias{QRMonRescale}
\title{Rescale data.}
\usage{
QRMonRescale(qrObj, timeAxisQ = TRUE, valueAxisQ = FALSE)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{timeAxisQ}{Should the data be rescaled along the time axis?}

\item{valueAxisQ}{Should the data be rescaled along the value axis?}
}
\value{
A QRMon object.
}
\description{
Rescale the data along axes specification.
}
\details{
The rescaled data replaces \code{qrObj$Data}.
}
