% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRMon.R
\name{QRMonSimulate}
\alias{QRMonSimulate}
\title{Simulate data points.}
\usage{
QRMonSimulate(qrObj, n = 100, points = NULL)
}
\arguments{
\item{qrObj}{An QRMon object.}

\item{n}{Number of simulated points.
Ignored if \code{points} is not NULL.}

\item{points}{A numerical vector of regressor points.
If NULL \code{n} is used.}
}
\value{
A QRMon object.
}
\description{
Simulate data points based on regression quantiles.
}
\details{
At each simulation regressor point:
(1) the corresponding regression quantiles values are found;
(2) then a interval of two regression probabilities is randomly picked;
(3) random point is generated between interval's quantiles.
Uniform distribution is used for the random point generation.
The obtained data frame with the simulated points is
assigned to \code{qrObj$Value}.
}
